% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniformance_class.R
\name{Uniformance}
\alias{Uniformance}
\title{Uniformance Class}
\description{
Creates an instance of the Uniformance class. This class acts as a wrapper for a 32-bit subprocess that accesses the honeywell Process Historian via honeywell's .NETFramework dll, phdapinet.dll.

The parameters (tag names, start, etc) are stored within the class, when requests for data the process is called with data being passed via an XML format.

Methods can be accessed for a class by creating an instance of the class and then using the following format "instance"$"Method"
}
\examples{
example <- Uniformance$new("MALSHW1")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Uniformance-new}{\code{Uniformance$new()}}
\item \href{#method-Uniformance-show_parameters}{\code{Uniformance$show_parameters()}}
\item \href{#method-Uniformance-check_tag}{\code{Uniformance$check_tag()}}
\item \href{#method-Uniformance-add_tag}{\code{Uniformance$add_tag()}}
\item \href{#method-Uniformance-show_taglist}{\code{Uniformance$show_taglist()}}
\item \href{#method-Uniformance-clear_taglist}{\code{Uniformance$clear_taglist()}}
\item \href{#method-Uniformance-remove_tag}{\code{Uniformance$remove_tag()}}
\item \href{#method-Uniformance-set_SampleFrequency}{\code{Uniformance$set_SampleFrequency()}}
\item \href{#method-Uniformance-SampleFrequency}{\code{Uniformance$SampleFrequency()}}
\item \href{#method-Uniformance-set_UseSampleFrequency}{\code{Uniformance$set_UseSampleFrequency()}}
\item \href{#method-Uniformance-UseSampleFrequency}{\code{Uniformance$UseSampleFrequency()}}
\item \href{#method-Uniformance-set_SampleFrequencyType}{\code{Uniformance$set_SampleFrequencyType()}}
\item \href{#method-Uniformance-SampleFrequencyType}{\code{Uniformance$SampleFrequencyType()}}
\item \href{#method-Uniformance-set_ReductionFrequency}{\code{Uniformance$set_ReductionFrequency()}}
\item \href{#method-Uniformance-ReductionFrequency}{\code{Uniformance$ReductionFrequency()}}
\item \href{#method-Uniformance-set_ReductionType}{\code{Uniformance$set_ReductionType()}}
\item \href{#method-Uniformance-ReductionType}{\code{Uniformance$ReductionType()}}
\item \href{#method-Uniformance-set_ReductionOffset}{\code{Uniformance$set_ReductionOffset()}}
\item \href{#method-Uniformance-ReductionOffset}{\code{Uniformance$ReductionOffset()}}
\item \href{#method-Uniformance-set_Startime}{\code{Uniformance$set_Startime()}}
\item \href{#method-Uniformance-set_Endtime}{\code{Uniformance$set_Endtime()}}
\item \href{#method-Uniformance-Startime}{\code{Uniformance$Startime()}}
\item \href{#method-Uniformance-Endtime}{\code{Uniformance$Endtime()}}
\item \href{#method-Uniformance-get_results}{\code{Uniformance$get_results()}}
\item \href{#method-Uniformance-clone}{\code{Uniformance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-new"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-new}{}}}
\subsection{Method \code{new()}}{
Initialises instances of the uniformance class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$new(hostname = NA, username = "", password = "", port = 3000)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-show_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-show_parameters}{}}}
\subsection{Method \code{show_parameters()}}{
Return a dataframe of all current parameters the first dataframe contains server parameters and the second contains details related to data to be acquired.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$show_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-check_tag"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-check_tag}{}}}
\subsection{Method \code{check_tag()}}{
Checks if tag exists on server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$check_tag(tag_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag_name}}{'A.RL_AI7361.BATCH'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-add_tag"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-add_tag}{}}}
\subsection{Method \code{add_tag()}}{
Checks if tag exists on server and adds to a list of tags that will be grabbed by get_data
returns current taglist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$add_tag(tag_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag_name}}{'A.RL_AI7361.BATCH'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-show_taglist"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-show_taglist}{}}}
\subsection{Method \code{show_taglist()}}{
Shows taglist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$show_taglist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-clear_taglist"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-clear_taglist}{}}}
\subsection{Method \code{clear_taglist()}}{
removes all tags from taglist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$clear_taglist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-remove_tag"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-remove_tag}{}}}
\subsection{Method \code{remove_tag()}}{
Remove a given tag
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$remove_tag(tag_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag_name}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-set_SampleFrequency"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-set_SampleFrequency}{}}}
\subsection{Method \code{set_SampleFrequency()}}{
The SampleFrequency property identifies the re-sample frequency, in seconds, used when invoking the get_results method.
Sets the sample Frequency, value is in seconds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$set_SampleFrequency(SampleFrequency)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SampleFrequency}}{Default value is 0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-SampleFrequency"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-SampleFrequency}{}}}
\subsection{Method \code{SampleFrequency()}}{
See current Sample Frequency
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$SampleFrequency()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-set_UseSampleFrequency"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-set_UseSampleFrequency}{}}}
\subsection{Method \code{set_UseSampleFrequency()}}{
Sets the use sample frequency flag
If True: The effective Reduction Frequency is the SampleFrequency.
If False: The ReductionFrequency property specifies the Reduction Frequency.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$set_UseSampleFrequency(useSampleFrequency)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{useSampleFrequency}}{Default value is FALSE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-UseSampleFrequency"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-UseSampleFrequency}{}}}
\subsection{Method \code{UseSampleFrequency()}}{
Sees the use sample frequency flag
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$UseSampleFrequency()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-set_SampleFrequencyType"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-set_SampleFrequencyType}{}}}
\subsection{Method \code{set_SampleFrequencyType()}}{
Sets the Sample Frequency Type Accepts:  Snapshot, Average, Resampled, Raw, InterpolatedRaw, Defaults to Raw
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$set_SampleFrequencyType(SampleType)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SampleType}}{Average
The value is the average from the data around the sample time. The time-weighted value for the raw samples are averaged so that the data value returned is an average for the value at the timestamp.

Resampled
The value is resampled from the raw data based on the resample method configured for the tag(s). Each tag may have a different resample method.

Interpolate Raw
Similar to the retrieval of raw data except that no data of -1 confidence is returned. When data of -1 confidence is encountered, the data for this point is interpolated from the surrounding data points and a confidence of 0 is applied to it. Unlike raw data, reduction values and unit conversion can be done using this method.

Raw
The value is the raw data. If more than one tag is being requested, the resulting data set will not be orthogonal as each tag will return a different number of values. No reduction values or unit conversion can be done when retrieving raw data.

Snapshot
The value is the interpolation between the raw data values at the sample time.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-SampleFrequencyType"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-SampleFrequencyType}{}}}
\subsection{Method \code{SampleFrequencyType()}}{
Sees the use Sample Frequency Type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$SampleFrequencyType()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-set_ReductionFrequency"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-set_ReductionFrequency}{}}}
\subsection{Method \code{set_ReductionFrequency()}}{
The ReductionFrequency property specifies the reduction frequency for any tag reductions being fetched. When using the ReductionFrequency property, the UseSampleFrequency flag must be False.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$set_ReductionFrequency(ReductionFrequency)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ReductionFrequency}}{Default value is 60}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-ReductionFrequency"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-ReductionFrequency}{}}}
\subsection{Method \code{ReductionFrequency()}}{
Sees the use Reduction Frequency
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$ReductionFrequency()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-set_ReductionType"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-set_ReductionType}{}}}
\subsection{Method \code{set_ReductionType()}}{
Sets the Reduction Type Accepts:  Snapshot, Average, Resampled, Raw, InterpolatedRaw,
"StandardDeviation", "RegressionSlope", "RegressionConstant",
"RegressionDeviation","First", "Last"
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$set_ReductionType(ReductionType)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ReductionType}}{Default value is FALSE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-ReductionType"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-ReductionType}{}}}
\subsection{Method \code{ReductionType()}}{
Sees the Reduction Type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$ReductionType()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-set_ReductionOffset"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-set_ReductionOffset}{}}}
\subsection{Method \code{set_ReductionOffset()}}{
Sets the Reduction Offset.
 The ReductionOffset property identifies the offset applied when retrieving reductions.
 Accepts: After, Around, Before
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$set_ReductionOffset(set_ReductionOffset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_ReductionOffset}}{Default value is Around}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-ReductionOffset"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-ReductionOffset}{}}}
\subsection{Method \code{ReductionOffset()}}{
Sees the Reduction Type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$ReductionOffset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-set_Startime"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-set_Startime}{}}}
\subsection{Method \code{set_Startime()}}{
Sets the starttime to be passed to the results function.
This takes either a date in either 'DD/MM/YYYY HH:mm:ss' or NOW-1d (d=day, w=week, y=year)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$set_Startime(starttime)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{starttime}}{'DD/MM/YYYY HH:mm:ss' format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-set_Endtime"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-set_Endtime}{}}}
\subsection{Method \code{set_Endtime()}}{
Sets the Endtime to be passed to the results function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$set_Endtime(endtime)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endtime}}{'DD/MM/YYYY HH:mm:ss' format
This takes either a date in either 'DD/MM/YYYY HH:mm:ss' or NOW-1d (d=day, w=week, y=year)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-Startime"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-Startime}{}}}
\subsection{Method \code{Startime()}}{
See current Starttime
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$Startime()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-Endtime"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-Endtime}{}}}
\subsection{Method \code{Endtime()}}{
See current Endtime
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$Endtime()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endtime}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-get_results}{}}}
\subsection{Method \code{get_results()}}{
Returns a list of dataframes with current tags and parameters. 

This method requires that tags have been added via add_tag().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
