% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniformance_class.R
\name{Uniformance}
\alias{Uniformance}
\title{Uniformance Class}
\description{
Creates an instance of the Uniformance class, This class acts as a wrapper for the phdapinet.dll which is used to access honeywell PHD servers/
The class holds the server data and methods for grabbing data.

Methods can be accessed for a class by creating an instance of the class and then using the following format "instance"$"Method"
}
\examples{
example <- Uniformance$new("MALSHW1")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Uniformance-new}{\code{Uniformance$new()}}
\item \href{#method-Uniformance-StartTime}{\code{Uniformance$StartTime()}}
\item \href{#method-Uniformance-Set_StartTime}{\code{Uniformance$Set_StartTime()}}
\item \href{#method-Uniformance-EndTime}{\code{Uniformance$EndTime()}}
\item \href{#method-Uniformance-Set_EndTime}{\code{Uniformance$Set_EndTime()}}
\item \href{#method-Uniformance-Sample_Frequency}{\code{Uniformance$Sample_Frequency()}}
\item \href{#method-Uniformance-Set_Sample_Frequency}{\code{Uniformance$Set_Sample_Frequency()}}
\item \href{#method-Uniformance-add_tag}{\code{Uniformance$add_tag()}}
\item \href{#method-Uniformance-list_tags}{\code{Uniformance$list_tags()}}
\item \href{#method-Uniformance-remove_all_tags}{\code{Uniformance$remove_all_tags()}}
\item \href{#method-Uniformance-get_results}{\code{Uniformance$get_results()}}
\item \href{#method-Uniformance-get_results_split}{\code{Uniformance$get_results_split()}}
\item \href{#method-Uniformance-clone}{\code{Uniformance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-new"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-new}{}}}
\subsection{Method \code{new()}}{
Initalises instiances of the uniformance class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$new(host = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host}}{This is the server where the tags you are interested in are, typically 'MALSHW1' or 'MBTSHW1'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-StartTime"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-StartTime}{}}}
\subsection{Method \code{StartTime()}}{
Returns current Start time for data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$StartTime()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-Set_StartTime"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-Set_StartTime}{}}}
\subsection{Method \code{Set_StartTime()}}{
Set start time for data
Defaults to Now-1hour
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$Set_StartTime(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{'DD/MM/YYYY HH:mm:ss' format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-EndTime"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-EndTime}{}}}
\subsection{Method \code{EndTime()}}{
Returns current end time for data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$EndTime()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-Set_EndTime"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-Set_EndTime}{}}}
\subsection{Method \code{Set_EndTime()}}{
Set end time for data
Defaults to Now
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$Set_EndTime(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{'DD/MM/YYYY HH:mm:ss' format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-Sample_Frequency"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-Sample_Frequency}{}}}
\subsection{Method \code{Sample_Frequency()}}{
Returns current Sample Frequency for data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$Sample_Frequency()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-Set_Sample_Frequency"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-Set_Sample_Frequency}{}}}
\subsection{Method \code{Set_Sample_Frequency()}}{
Set the Sample Frequency for data for data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$Set_Sample_Frequency(frequency)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{frequency}}{Value should be in seconds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-add_tag"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-add_tag}{}}}
\subsection{Method \code{add_tag()}}{
add tags to tag list that will be checked, tags should be provided as a string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$add_tag(tag_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag_name}}{'A.RL_AI7361.BATCH'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-list_tags"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-list_tags}{}}}
\subsection{Method \code{list_tags()}}{
Returns a list of all the current tags
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$list_tags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-remove_all_tags"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-remove_all_tags}{}}}
\subsection{Method \code{remove_all_tags()}}{
Removes all tags
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$remove_all_tags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-get_results}{}}}
\subsection{Method \code{get_results()}}{
Returns a dataframe of all data using currently set parameters and tags
This method requires that tags have been added via add_tag()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-get_results_split"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-get_results_split}{}}}
\subsection{Method \code{get_results_split()}}{
Returns a dataframe of all data using currently set parameters and tags
This method requires that tags have been added via add_tag()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$get_results_split()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
