% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniformance_class.R
\name{Uniformance}
\alias{Uniformance}
\title{Uniformance Class}
\description{
Creates an instance of the Uniformance class. This class acts as a wrapper for a 32-bit subprocess that accesses the honeywell Process Historian via honeywell's .NETFramework dll, phdapinet.dll.

The parameters (tag names, start, etc) are stored within the class, when requests for data the process is called with data being passed via an XML format.

Methods can be accessed for a class by creating an instance of the class and then using the following format "instance"$"Method"
}
\examples{
example <- Uniformance$new("MALSHW1")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Uniformance-new}{\code{Uniformance$new()}}
\item \href{#method-Uniformance-add_tag}{\code{Uniformance$add_tag()}}
\item \href{#method-Uniformance-show_taglist}{\code{Uniformance$show_taglist()}}
\item \href{#method-Uniformance-clear_taglist}{\code{Uniformance$clear_taglist()}}
\item \href{#method-Uniformance-remove_tag}{\code{Uniformance$remove_tag()}}
\item \href{#method-Uniformance-set_startime}{\code{Uniformance$set_startime()}}
\item \href{#method-Uniformance-set_endtime}{\code{Uniformance$set_endtime()}}
\item \href{#method-Uniformance-set_samplefrequency}{\code{Uniformance$set_samplefrequency()}}
\item \href{#method-Uniformance-startime}{\code{Uniformance$startime()}}
\item \href{#method-Uniformance-endtime}{\code{Uniformance$endtime()}}
\item \href{#method-Uniformance-samplefrequency}{\code{Uniformance$samplefrequency()}}
\item \href{#method-Uniformance-get_results}{\code{Uniformance$get_results()}}
\item \href{#method-Uniformance-clone}{\code{Uniformance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-new"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-new}{}}}
\subsection{Method \code{new()}}{
Initialises instances of the uniformance class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$new(hostname = NA, username = "", password = "", port = 3000)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-add_tag"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-add_tag}{}}}
\subsection{Method \code{add_tag()}}{
Checks if tag exists on server and adds to a list of tags that will be grabbed by get_data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$add_tag(tag_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag_name}}{'A.RL_AI7361.BATCH'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-show_taglist"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-show_taglist}{}}}
\subsection{Method \code{show_taglist()}}{
Shows taglist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$show_taglist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-clear_taglist"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-clear_taglist}{}}}
\subsection{Method \code{clear_taglist()}}{
Clears taglist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$clear_taglist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-remove_tag"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-remove_tag}{}}}
\subsection{Method \code{remove_tag()}}{
Remove a given tag
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$remove_tag(tag_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag_name}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-set_startime"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-set_startime}{}}}
\subsection{Method \code{set_startime()}}{
Sets the starttime to be passed to the results function.
This takes either a date in either 'DD/MM/YYYY HH:mm:ss' or NOW-1d (d=day, w=week, y=year)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$set_startime(starttime)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{starttime}}{'DD/MM/YYYY HH:mm:ss' format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-set_endtime"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-set_endtime}{}}}
\subsection{Method \code{set_endtime()}}{
Sets the Endtime to be passed to the results function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$set_endtime(endtime)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endtime}}{'DD/MM/YYYY HH:mm:ss' format
This takes either a date in either 'DD/MM/YYYY HH:mm:ss' or NOW-1d (d=day, w=week, y=year)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-set_samplefrequency"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-set_samplefrequency}{}}}
\subsection{Method \code{set_samplefrequency()}}{
Sets the sample Frequency, value is in seconds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$set_samplefrequency(samplefrequency)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samplefrequency}}{Default value is 0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-startime"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-startime}{}}}
\subsection{Method \code{startime()}}{
See current Starttime
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$startime()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-endtime"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-endtime}{}}}
\subsection{Method \code{endtime()}}{
See current Endtime
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$endtime()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endtime}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-samplefrequency"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-samplefrequency}{}}}
\subsection{Method \code{samplefrequency()}}{
See current Sample Frequency
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$samplefrequency()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-get_results}{}}}
\subsection{Method \code{get_results()}}{
Returns a list of dataframes with current tags and parameters. 

This method requires that tags have been added via add_tag().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniformance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Uniformance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniformance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
